requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Slack Integration
slack-bolt==1.18.1
slack-sdk==3.26.2

# Google Gemini
google-generativeai==0.3.2

# Vector Store & Embeddings
faiss-cpu>=1.12.0
numpy>=1.26.0

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Production
gunicorn==21.2.0
